Set Up Basic CheckPoint Management System
Just A System Of A Manager and A Prototype CheckPoint With Basic Data and Visualization Through UnityEditor, evolving in a Custom Checkpoint and Triggers Editor.
Useful (I hope) for Managing Areas Which Trigger Events such as Audio Cues, Story Events or Environmental Behaviours (Still to be implemented via a common Interface).
ValidLayerAgents, EventTags and collision Areas will allow for different kind of interactions with other systems.